enhance the flexibility of the test paramter "num-errors"
in some conditions, num-errors can't be defined exactly. For
example, we expect the some error happens 10 times. Maybe
it is true on machine A but on machine B it is not, say, 8.
If we define the num-errors is 10, on machine B the test will
not end until we break it, it is because the mcelog is still
waiting the left errors.

Signed-off-by: Chen Gong <gong.chen@linux.intel.com>