mcelog: Implement dmi decoding for UEFI
When mcelog is executed on the UEFI system, mcelog shows
the following error message:
    mcelog: Cannot find SMBIOS DMI tables

While with legacy x86 (non-EFI) systems we can scan the
range 0xf0000 to 0xfffff for the entry point of SMBIOS
structure table, we can not do the same thing with
EFI-based systems.

This patch is used for getting the SMBIOS structure tables
on x86 UEFI systems.

Signed-off-by: Hidetoshi Seto <seto.hidetoshi@jp.fujitsu.com>
Signed-off-by: Jin Dongming <jin.dongming@np.css.fujitsu.com>
Signed-off-by: Andi Kleen <ak@linux.intel.com>