Add --debug-numerrors
Based on a earlier patch by Chen Gong.

in some conditions, num-errors can't be defined exactly. For
example, we expect the some error happens 10 times. Maybe
it is true on machine A but on machine B it is not, say, 8.
If we define the num-errors is 10, on machine B the test will
not end until we break it, it is because the mcelog is still
waiting the left errors.

[AK: Changed original patch to use a new option]

Signed-off-by: Andi Kleen <ak@linux.intel.com>