Fix unsupported test at start of mcelog
This makes testing in an initscript and an rpm install easier.  In the case of
the rpm install I can use the postinstall section to test before installing
the hourly cron job.

It doesn't matter whether or not we're in daemon mode; either the CPU is
supported or it isn't.  If it isn't supported bail out with a warning.

Also fix up the warning to output to stderr.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>
Cc: Borislav Petkov <borislav.petkov@amd.com>
Signed-off-by: Andi Kleen <ak@linux.intel.com>