Add support for Sandy Bridge extended error logging
Sandy Bridge -EP processors have a mode where extra logging information
for corrected errors is provided in the memory controller bank allowing
us to identify the DIMM that was the source of the error.

This mode is disabled in hardware by default (so that legacy systems
are not confused by seeing two banks report each corrected error),
but can be turned on by setting a bit in MSR 0x17f.

Add code to turn this mode on in --daemon mode (with a --no-imc-log
option to turn it off if there are problems, or to --ascii decode logs
from systems where it is not enabled).

Signed-off-by: Tony Luck <tony.luck@intel.com>
Signed-off-by: Andi Kleen <ak@linux.intel.com>