Modify --daemon logic to allow syslog-only output
At some point, the --daemon option was modified to force the usage of
--logfile.  Either the user sets it or the daemon uses a default value.  I
suspect this is to ensure that the daemon doesn't start without a logging
output.  This is a good idea, but the implementation also makes it impossible
to run the daemon in a syslog-only mode.

This patch changes the behavior so that, instead of --daemon implying a default
--logfile, --daemon without an output option implies the default --logfile.

With this change, "mcelog --daemon --syslog" no longer has the confusing
behavior of opening the default logfile.  The other configuration paths should
be unchanged.