Skip Hardware error prefix
The kernel prefixes the lines that we want to parse with both the
module name ("mce") and the HW_ERR indicator ("[Hardware Error]") both
suffixed with a ":".  Allow users to paste the raw kernel output from
serial console or dmesg(8) output by making mcelog skip over this to
get to the meat of the message.

Signed-off-by: Tony Luck <tony.luck@intel.com>
Signed-off-by: Andi Kleen <ak@linux.intel.com>