Fix error count during the last agetime
The error count during the last 24h on channel any DIMM any doesn't match
the sum of the errors on all channels and DIMMs. The disparity is because
the statistics are only updated when an error occurs. They don't age with
time.

Fix it by calling bucket_age() when reading them out.

Signed-off-by: Rui Wang <rui.y.wang@intel.com>
Signed-off-by: Andi Kleen <ak@linux.intel.com>