mcelog: is_cpu_supported() error message must be printed Eprintf
SYSERRprintf outputs a ": Success" message so the error message looks like

mcelog: ERROR: AMD Processor family 21: mcelog does not support this
processor.  Please use the edac_mce_amd module instead.
: Success

which is confusing for end-users.

I changed this to do Eprintf which keeps the error return to userspace
but drops the ": Success" line.

Signed-off-by: Prarit Bhargava <prarit@redhat.com>
Signed-off-by: Andi Kleen <ak@linux.intel.com>