Fix structure offset checks
Mcelog tries to cope with changes in the size of the "struct mce"
that is read from /dev/mcelog since the kernel may not be updated
in sync with the utility. To achieve this it makes an ioctl call to
the kernel to check the size of its version of the structure. There
are two cases:

1) If the kernel has a larger structure, then mcelog skips the extra
fields because it doesn't know what they are.

2) If the kernel has a smaller structure, then mcelog must avoid using
any fields beyond the end of the kernel structure.

The test for this second case looks like thi:

	if (recordlen >= offsetof(struct mce, cpuid) && m->mcgcap)
		n += Wprintf("MCGCAP %llx ", m->mcgcap);

but the test is incorrect. We should check for greater than, not
greater than or equal.

This was noticed on a system with a kernel that didn't have support
for the ppin field, but mcelog was printing:

	PPIN ec00000000010090

(the value printed was the "status" field from the next "mce" structure).

Fix all the tests of this form to use ">" instead of ">=".

Signed-off-by: Tony Luck <tony.luck@intel.com>
Signed-off-by: Andi Kleen <ak@linux.intel.com>