mcelog: Fix "--ascii" parsing to cope with change in kernel output since v4.10
Linux kernel now includes this change:

 commit cd9c57cad3fe ("x86/MCE: Dump MCE to dmesg if no consumers")

which has this fragment:

-       pr_emerg(HW_ERR "CPU %d: Machine Check Exception: %Lx Bank %d: %016Lx\n",
-              m->extcpu, m->mcgstatus, m->bank, m->status);
+       pr_emerg(HW_ERR "CPU %d: Machine Check%s: %Lx Bank %d: %016Lx\n",
+                m->extcpu,
+                (m->mcgstatus & MCG_STATUS_MCIP ? " Exception" : ""),
+                m->mcgstatus, m->bank, m->status);

that switched from always printing the word "Exception" to only
doing so for errors that were logged during aan exception (i.e. when
IA32_MSR_MCG_STATUS.MCIP is set to one).  This is helpful to human readers
of console logs to distinguish between errors that caused machine checks
from errors that were found later polling machine check banks.  But it
broke the "--ascii" parsing code in mcelog(8) :-(

The change went upstream 15 months ago in v4.10. So the horse has
left the stable and is over the hills and far away.

Fix the parsing code to treat the word "Exception" as optional.

Signed-off-by: Tony Luck <tony.luck@intel.com>
Signed-off-by: Andi Kleen <ak@linux.intel.com>