mcelog: Improve decoding for APEI reported errors
The Linux kernel converts memory errors reported to it by APEI
into a roughly equivalent error reported in a machine check bank.
It used to make these appear as if they were reported in bank 1,
but this caused confusion to h/w folks because memory errors are
not logged in bank 1 on most CPU models.

The kernel switched to using bank -1 starting in v4.8 with commit
  b2de43605410 ("x86/mce: Do not use bank 1 for APEI generated error logs")

Since the "bank" field in the "mce" structure is of type "u8", this
shows up as 255 to mcelog.  Reports look like:

  CPU 0 Undecoded extended event ff

Update the extended_bankname() function to know about this and give
the more helpful:

  CPU 0 ACPI/APEI reported error

Signed-off-by: Tony Luck <tony.luck@intel.com>
Signed-off-by: Andi Kleen <ak@linux.intel.com>