mcelog: Use 'num-errors' to specify the number of mce records to be injected.
In the file tests/page/inject, the default number of mce records to
be injected is 6. Some test cases may need more then 6 mce records
to be injected for testing, e.g., 'page-error-counter-replacement'.

So use 'num-errors' in the configuration file to specify the number
of mce records to be injected. And inject mce records and run mcelog
in parallel, so that the mce records can be consumed by mcelog in
time to avoid mce records overflow.

Signed-off-by: Qiuxu Zhuo <qiuxu.zhuo@intel.com>
Signed-off-by: Andi Kleen <ak@linux.intel.com>