mcelog: Fix issues with "yellow" cache offlining
An Ice Lake system reported a cache error with this signature in MCi_STATUS
0x8c4238c000071152. Decoding the low 16 bits (MCACOD) gives:

	0001		0001		0101	0010
	   Filter	   Cache	RRRR	TTLL

TT = 00 means the error occurred on an instruction transaction
LL = 10 means in cache level 2 (counting from zero)

Two problems in mcelog handling this error signature:

1) In the decode of a compound error code the "LL" field refers to
the cache hierarchy counting from zero. So a typical modern CPU
with separate L1 Instruction and Data caches count these as level 0.
The mid-level cache is level 1. The last level cache is level 2.

But Linux /sys topology counts from 1. So it shows:
$ grep . /sys/devices/system/cpu/cpu0/cache/index*/level
/sys/devices/system/cpu/cpu0/cache/index0/level:1
/sys/devices/system/cpu/cpu0/cache/index1/level:1
/sys/devices/system/cpu/cpu0/cache/index2/level:2
/sys/devices/system/cpu/cpu0/cache/index3/level:3

As a result of this mismatch mcelog thinks the only affected
CPUs are those sharing the L2 cache. So only the HT threads
sharing a core are listed in the AFFECTED_CPUS environment
variable passed to the trigger script.

FIX: Adjust the cache level by adding one to match both the
Linux numbering and also the Intel SDM naming of cache levels.
I.e. 1, 2, 3 rather than 0, 1, 2.

2) The cache_to_cpus() expects a 1:1 match of the transaction type
to the type of the cache specified in the LL field. This fails because
there is no "Instruction" cache at level 2. Mcelog prints:

	Cannot find sysfs cache for CPU 28
	Offlining CPU unknown due to cache error threshold

FIX: Update the check to first check for an exact match of both level and
type as before, but also allow a match at the right level with a "UNIFIED"
cache).

Signed-off-by: Tony Luck <tony.luck@intel.com>
Signed-off-by: Andi Kleen <ak@linux.intel.com>