mcelog: Makefile: Only touch cputype.h if needed to create it
The "touch cputype.h" was added to the .depend rule to create the file
if it did not exist so that the following $(CC) command doesn't give
errors if this generated file is missing.

But touching it when it does exist results in rebuilding all the .c
files that include it.

Only touch cputype.h if it does NOT exist.

Signed-off-by: Tony Luck <tony.luck@intel.com>
Signed-off-by: Andi Kleen <ak@linux.intel.com>