mcelog: Don't overwrite model number when lookup fails
When converting to table driven generation of mappings between model
numbers and names etc. the new code overwrote the model number that
was found in the "struct mce" with the return value from
lookup_intel_cputype(). If that function succeeded, everything worked
well. But if it failed, then mcelog would print:

 mcelog: Unknown Intel CPU type family 6 model 4294967295

That huge number is the unsigned 32-bit value for "-1", the
error return value from lookup_intel_cputype().

Save the return from lookup_intel_cputype() in a separate variable.

Fixes: 9aee26472b2b ("mcelog: Add CPU model numbers to table and generate switch function")
Signed-off-by: Tony Luck <tony.luck@intel.com>