mcelog: Add kernel CONFIG message if /dev/mcelog isn't found
Linux kernel made support for /dev/mcelog optional in v4.12. Most
distributions built with CONFIG_X86_MCELOG_LEGACY=y for a while.
But some have started to drop it.

When /dev/mcelog cannot be opened, provide an extra message to tell
users to check if their kernel is built with this option.

Signed-off-by: Tony Luck <tony.luck@intel.com>