mcelog: Add a --binary option for reading records saved to pstore
The Linux kernel can be configured to save fatal error records to
persistent storage with the pstore file system. These are a raw
copy of "struct mce".

Add an option to skip the ioctl() calls that determine the record
size so that mcelog will decode a binary file given as argument.

Signed-off-by: Tony Luck <tony.luck@intel.com>